requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
reportlab==4.0.7
icalendar==5.0.11
geopy==2.4.1
googlemaps==4.10.0
pydantic==2.5.0
python-dateutil==2.8.2

